Add ParkingSpotController and integrate data table columns with filters for status and municipality
